Clean up and move enrichment into Extensions.Logging
